implemented inheretance from BaseTest
